pyproject.toml
[project]
name = "chatbot-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.79.0",
    "dill>=0.4.1",
    "janome>=0.5.0",
    "japanize-matplotlib>=1.1.3",
    "jupyterlab>=4.5.3",
    "langchain>=1.2.9",
    "langchain-anthropic>=1.3.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "openai>=2.17.0",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
]
